+++ libs/fd.h
@@ -1,4 +1,12 @@
 // SPDX-License-Identifier: (LGPL-2.1 OR BSD-2-Clause)
 #pragma once
 
+#ifdef __cplusplus
+extern "C" {
+#endif
+
 int setnonblock(int fd);
+
+#ifdef __cplusplus
+}
+#endif