+++ libs/mac.h
@@ -1,3 +1,12 @@
 // SPDX-License-Identifier: (LGPL-2.1 OR BSD-2-Clause)
 #pragma once
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
 int strtomacaddr(unsigned char *mac, char *str);
+
+#ifdef __cplusplus
+}
+#endif