+++ libs/pcie_helpers.h
@@ -3,6 +3,10 @@
 #pragma once
 #include <stdint.h>
 
+#ifdef __cplusplus
+extern "C" {
+#endif
+
 /**
  * PCI-Compatible Configuration Registers
  *
@@ -123,3 +127,7 @@ const char *pci_cs_type_name(uint8_t header_type, char *buf, size_t buf_sz);
 void pci_cs_print_common(struct pci_cs_hdr_common *c);
 void pci_cs_print_type0(struct pci_cs_hdr_type0 *t);
 void pci_cs_print_type1(struct pci_cs_hdr_type1 *t);
+
+#ifdef __cplusplus
+}
+#endif