+++ libs/str.h
@@ -1,6 +1,10 @@
 // SPDX-License-Identifier: (LGPL-2.1 OR BSD-2-Clause)
 #pragma once
 
+#ifdef __cplusplus
+extern "C" {
+#endif
+
 void *memcpy_c(void *dest, const void *src, size_t n);
 char *strcaseswap(char *str, ssize_t len);
 char *vstrjoint(char *dst, const char *fmt, ...);
@@ -9,3 +13,7 @@ const char *vstrcat_r(char buf[], int nstr, ...);
 
 unsigned long str2size(const char *str);
 unsigned long str2nsecs(const char *str);
+
+#ifdef __cplusplus
+}
+#endif