+++ libs/syscall_helpers.h
@@ -3,5 +3,13 @@
 #pragma once
 #include <sys/types.h>
 
+#ifdef __cplusplus
+extern "C" {
+#endif
+
 int create_file(const char *filename, char ch, size_t bytes);
 off_t file_seek_end(const char *filename);
+
+#ifdef __cplusplus
+}
+#endif